+++ src/serialization/Serializable.ts
@@ -1,4 +1,5 @@
 import {DataInput, DataOutput} from './Data';
+import {PortableWriter, PortableReader} from './Portable';
 export interface IdentifiedDataSerializable {
     readData(input: DataInput): any;
     writeData(output: DataOutput): void;
@@ -8,4 +9,15 @@ export interface IdentifiedDataSerializable {
 
 export interface IdentifiedDataSerializableFactory {
     create(type: number): IdentifiedDataSerializable;
+}
+
+export interface Portable {
+    getFactoryId(): number;
+    getClassId(): number;
+    writePortable(writer: PortableWriter): void;
+    readPortable(reader: PortableReader): void;
+}
+
+export interface PortableFactory {
+    create(classId: number): Portable;
 }
